add "drop" arg to main.  sw() calls upon drop.py.
"exceptions" arg removed from main's list of args.  "drop" added to the
list.  At the program's head, added "import drop".  And added the lines
just previous so as to implement import drop.  In the sw() function
adjusted the code that establishes filtr3 so that lists from drop lists
called upon.