fixcsv() function updated.
fixcsv has been rewritten.  The obsolete function depended on the
DESCRIPTION column of a SW csv file ALWAYS being 2nd to last.  And it
depended on the DESCRIPTION field only having 0 or 1 commas.  The new
fixcsv file has no restriction on where the DESCRIPTION column is
located.  And the DESCRIPTION field of a part can have any number of
commas.