Avoid files that begin with ~.  Implement timezone.
Temp files that begin with ~, like ~$086987_sl.xlsx, cause the program
to crash.  This fixed in the gatherBOMs function.  Use of a new
variable named timezone extablished.  This ensures that a proper date &
time is established if the bomcheck program is run on a remote server
outside of Dekker's timezone.