To date: Except .txt, Ignore "in", Ignore lower case.
Changes to date for v1.0.16:
1.  A file name with the extension "txt" can now be read in as a
csv file (or at least try to read it in).
2.  In a SW BOM table, if a value in the LENGTH column is a text
value like '3.52"' or '3.52 inch', then ignore the '"'  or 'inch'
portion of the text.
3.  If a file name contains lower case letters, e.g.
dv1200d-kb4_sw.xlsx, then it wasn't accounted for in the BOM check.
This is no longer the case.  File names are now case insensitive.
4.  Instead of just one file name or a glob, '*', used as data input to
bomcheck, bomcheck not accepts a list of file names or globs.  This
change done to prepare it to be called upon with a gui interface
via PyQt.